+++ GeoFeatures/GFLineString.mm
@@ -95,7 +95,7 @@ namespace gf = geofeatures;
         auto size = _lineString.size();
         
         if (size == 0 || index > (size -1)) {
-            [NSException raise:NSRangeException format:@"Index %li is beyond bounds [0, %li].", (unsigned long) index, _lineString.size()];
+            [NSException raise:NSRangeException format:@"Index %li is beyond bounds [0, %li].", (unsigned long) index, (unsigned long) _lineString.size()];
         }
         //
         // Note: We use operator[] below because we've
@@ -130,7 +130,7 @@ namespace gf = geofeatures;
         auto size = _lineString.size();
 
         if (size == 0 || index > (size -1)) {
-            [NSException raise: NSRangeException format: @"Index %li is beyond bounds [0, %li].", (unsigned long) index, _lineString.size()];
+            [NSException raise: NSRangeException format: @"Index %li is beyond bounds [0, %li].", (unsigned long) index, (unsigned long) _lineString.size()];
         }
         //
         // Note: We use operator[] below because we've
@@ -193,7 +193,7 @@ namespace gf = geofeatures;
             [NSException raise: NSInvalidArgumentException format: @"aPoint can not be nil."];
         }
         if (index > _lineString.size()) {
-            [NSException raise: NSRangeException format: @"Index %li is beyond bounds [0, %li].", (unsigned long) index, _lineString.size()];
+            [NSException raise: NSRangeException format: @"Index %li is beyond bounds [0, %li].", (unsigned long) index, (unsigned long) _lineString.size()];
         }
         // TODO: Handle bad_alloc?
         _lineString.insert(_lineString.begin() + index, gf::Point([aPoint x], [aPoint y]));
@@ -206,7 +206,7 @@ namespace gf = geofeatures;
     - (void) removePointAtIndex: (NSUInteger) index {
 
         if (index >= _lineString.size()) {
-            [NSException raise: NSRangeException format: @"Index %li is beyond bounds [0, %li].", (unsigned long) index, _lineString.size()];
+            [NSException raise: NSRangeException format: @"Index %li is beyond bounds [0, %li].", (unsigned long) index, (unsigned long) _lineString.size()];
         }
         _lineString.erase(_lineString.begin() + index);
     }
